refactor(services): extract external-services and service-insights to legacy-services
