--- packages/kuma-gui/src/app/external-services/index.ts
@@ -1,27 +1,0 @@
-import { token } from '@kumahq/container'
-
-import locales from './locales/en-us/index.yaml'
-import { sources } from './sources'
-import type { ServiceDefinition } from '@kumahq/container'
-
-type Token = ReturnType<typeof token>
-
-export const services = (app: Record<string, Token>): ServiceDefinition[] => {
-  return [
-    [token('external-services.sources'), {
-      service: sources,
-      arguments: [
-        app.api,
-      ],
-      labels: [
-        app.sources,
-      ],
-    }],
-    [token('external-services.locales'), {
-      service: () => locales,
-      labels: [
-        app.enUs,
-      ],
-    }],
-  ]
-}
