--- packages/kuma-gui/src/app/external-services/locales/en-us/index.yaml
@@ -1,21 +1,0 @@
-external-services:
-  routes:
-    item:
-      title: "{name}"
-      breadcrumbs: ExternalServices
-      navigation:
-        external-service-detail-view: 'Overview'
-      overview: 'Overview'
-      config: 'YAML'
-      format: Format
-      formats:
-        universal: Universal
-        k8s: Kubernetes
-    items:
-      title: ExternalServices
-  detail:
-    about:
-      title: About this ExternalService
-    config: YAML
-  href:
-    docs: '{KUMA_DOCS_URL}/policies/external-services?{KUMA_UTM_QUERY_PARAMS}'
