+++ packages/kuma-gui/src/app/legacy-services/data/ExternalService.ts
