+++ packages/kuma-gui/src/app/legacy-services/data/index.ts
@@ -1,0 +1,2 @@
+export * from './ExternalService'
+export * from './ServiceInsight'
